Implement Code review
chore: move ChannelTypesSets interface to Channel.d.ts
chore: Implement selection for showing channel types
fix: implement try/catch for shouldShowHiddenChannel
chore: create ChannelTypeSets in utils
